Add code comments
