+++ upgrade/internal/crds.go
@@ -2,20 +2,23 @@ package internal
 
 import (
 	"fmt"
+	"log"
+
 	"github.com/numaproj/helm-charts/upgrade/common"
 )
 
 const crdsLocalPath = "crds/"
 
+// UpdateCRDFiles updates the CRD files in the local directory with the latest versions from the GitHub repository.
 func UpdateCRDFiles(numaflowVersion string) {
 	for fileName, path := range common.CRDFiles {
 		githubFileUrl := path + fileName
 		localFilePath := generateCRDsFilePath(fileName)
 		if err := updateFiles(localFilePath, githubFileUrl, numaflowVersion, false); err != nil {
-			println("Error updating CRD file:", fileName, "err:", err)
+			log.Println("Error updating CRD file:", fileName, "err:", err)
 			continue
 		}
-		fmt.Println("Successfully updated CRD file:", fileName, "...")
+		log.Println("Successfully updated CRD file:", fileName, "...")
 	}
 }
