+++ upgrade/internal/update-chart.go
@@ -2,7 +2,7 @@ package internal
 
 import (
 	"fmt"
-	"os"
+	"log"
 	"strconv"
 	"strings"
 
@@ -10,18 +10,17 @@ import (
 	"helm.sh/helm/v3/pkg/chartutil"
 )
 
+// UpdateChartFile updates the Chart.yaml file with the new version and appVersion.
 func UpdateChartFile(numaflowVersion string) {
 	chartFilePath := common.BaseDir + chartutil.ChartfileName
 	chart, err := chartutil.LoadChartfile(chartFilePath)
 	if err != nil {
-		fmt.Printf("Error loading chart file: %s\n", err)
-		os.Exit(1)
+		log.Fatalf("Error loading chart file: %s\n", err)
 	}
 
 	numaflowVersion = strings.TrimPrefix(numaflowVersion, "v")
 	if numaflowVersion == chart.AppVersion {
-		fmt.Println("Versions are identical. No update is necessary.")
-		os.Exit(1)
+		log.Fatalln("Versions are identical. No update is necessary.")
 	}
 	numaflowParts := strings.Split(numaflowVersion, ".")
 	appParts := strings.Split(chart.AppVersion, ".")
@@ -36,27 +35,23 @@ func UpdateChartFile(numaflowVersion string) {
 	case majorUpdate:
 		currentVersion, err := strconv.Atoi(currentParts[0])
 		if err != nil {
-			fmt.Printf("Error parsing current version: %s\n", err)
-			os.Exit(1)
+			log.Fatalf("Error parsing current version: %s\n", err)
 		}
 		newVersion = fmt.Sprintf("%d.0.0", currentVersion+1)
 	case minorUpdate:
 		currentVersion, err := strconv.Atoi(currentParts[1])
 		if err != nil {
-			fmt.Printf("Error parsing current version: %s\n", err)
-			os.Exit(1)
+			log.Fatalf("Error parsing current version: %s\n", err)
 		}
 		newVersion = fmt.Sprintf("%s.%d.0", currentParts[0], currentVersion+1)
 	case patchUpdate:
 		currentVersion, err := strconv.Atoi(currentParts[2])
 		if err != nil {
-			fmt.Printf("Error parsing current version: %s\n", err)
-			os.Exit(1)
+			log.Fatalf("Error parsing current version: %s\n", err)
 		}
 		newVersion = fmt.Sprintf("%s.%s.%d", currentParts[0], currentParts[1], currentVersion+1)
 	default:
-		fmt.Printf("No significant version change detected. No update to %s required.\n", numaflowVersion)
-		os.Exit(1)
+		log.Fatalf("No significant version change detected. No update to %s required.\n", numaflowVersion)
 	}
 
 	chart.Version = newVersion
@@ -64,9 +59,8 @@ func UpdateChartFile(numaflowVersion string) {
 
 	err = chartutil.SaveChartfile(chartFilePath, chart)
 	if err != nil {
-		fmt.Printf("Error saving chart file: %s\n", err)
-		os.Exit(1)
+		log.Fatalf("Error saving chart file: %s\n", err)
 	}
 
-	fmt.Printf("Updated version: %s and appVersion: %s in charts/numaflow/Chart.yaml\n", newVersion, numaflowVersion)
+	log.Printf("Updated version: %s and appVersion: %s in charts/numaflow/Chart.yaml", newVersion, numaflowVersion)
 }
