+++ upgrade/main.go
@@ -1,29 +1,32 @@
 package main
 
 import (
-	"fmt"
-	"github.com/numaproj/helm-charts/upgrade/internal"
 	"log"
 	"os"
+
+	"github.com/numaproj/helm-charts/upgrade/internal"
 )
 
 func main() {
-	numaflowVersion := os.Args[1]
+	numaflowVersion := os.Getenv("NUMAFLOW_VERSION")
 	if numaflowVersion == "" {
 		log.Fatalln("Numaflow version is required as a command line argument")
 	}
 
-	fmt.Println("Checking version existence in Numaflow repo...")
+	log.Println("Checking version existence in Numaflow repo...")
 	if exists, err := internal.IsVersionExists(numaflowVersion); !exists && err != nil {
 		log.Fatalln("Version check failed:", err)
 	}
 
-	fmt.Println("################### Updating Chart.yaml ###################")
+	log.Println("################### Updating Chart.yaml ###################")
 	internal.UpdateChartFile(numaflowVersion)
-	fmt.Println("\n################### Updating latest CRDs ###################")
+
+	log.Println("\n################### Updating latest CRDs ###################")
 	internal.UpdateCRDFiles(numaflowVersion)
-	fmt.Println("\n################### Updating latest data for RBAC ###################")
+
+	log.Println("\n################### Updating latest data for RBAC ###################")
 	internal.UpdateRBACFiles(numaflowVersion)
-	fmt.Println("\n################### Updating latest data for Service Account ###################")
+
+	log.Println("\n################### Updating latest data for Service Account ###################")
 	internal.UpdateServiceAccount(numaflowVersion)
 }
